Move download RC to a new file Move rc-params to a new file that can be accessed both on test-runner and download-rc.js
