+++ package.json
@@ -35,11 +35,11 @@
     "scripts": {
         "clean": "rimraf lib *.jar *.log",
         "compile": "tsc",
-        "test": "node scripts/test-util.js alltests",
-        "test:unit": "node scripts/test-util.js unit",
-        "test:integration": "node scripts/test-util.js integration",
+        "test": "node scripts/download-rc.js && node scripts/test-runner.js alltests",
+        "test:unit": "node scripts/test-runner.js unit",
+        "test:integration": "node scripts/download-rc.js && node scripts/test-runner.js integration",
         "validate-user-code": "tsc --build test/user_code/tsconfig.json",
-        "precoverage": "node scripts/test-util.js startrc",
+        "precoverage": "node scripts/download-rc.js && node scripts/test-runner.js startrc",
         "coverage": "rimraf coverage && nyc node_modules/mocha/bin/_mocha 'test/**/*.js' -- --reporter-options mochaFile=report.xml --reporter mocha-junit-reporter",
         "pregenerate-docs": "rimraf docs",
         "generate-docs": "typedoc --options typedoc.json",
