+++ scripts/rc-params.js
@@ -1,0 +1,28 @@
+'use strict';
+const HZ_VERSION = '4.1.1';
+const HZ_TEST_VERSION = '4.1.1';
+const HAZELCAST_TEST_VERSION = HZ_TEST_VERSION;
+const HAZELCAST_VERSION = HZ_VERSION;
+const HAZELCAST_ENTERPRISE_VERSION = HZ_VERSION;
+const HAZELCAST_ENTERPRISE_TEST_VERSION = HZ_VERSION;
+const HAZELCAST_RC_VERSION = '0.8-SNAPSHOT';
+const SNAPSHOT_REPO = 'https://oss.sonatype.org/content/repositories/snapshots';
+const RELEASE_REPO = 'http://repo1.maven.apache.org/maven2';
+const ENTERPRISE_RELEASE_REPO = 'https://repository.hazelcast.com/release/';
+const ENTERPRISE_SNAPSHOT_REPO = 'https://repository.hazelcast.com/snapshot/';
+const HAZELCAST_ENTERPRISE_KEY = process.env.HAZELCAST_ENTERPRISE_KEY ? process.env.HAZELCAST_ENTERPRISE_KEY : '';
+
+module.exports = {
+    HZ_VERSION: HZ_VERSION,
+    SNAPSHOT_REPO: SNAPSHOT_REPO,
+    ENTERPRISE_SNAPSHOT_REPO: ENTERPRISE_SNAPSHOT_REPO,
+    RELEASE_REPO: RELEASE_REPO,
+    ENTERPRISE_RELEASE_REPO: ENTERPRISE_RELEASE_REPO,
+    HAZELCAST_RC_VERSION: HAZELCAST_RC_VERSION,
+    HZ_TEST_VERSION: HZ_TEST_VERSION,
+    HAZELCAST_VERSION: HAZELCAST_VERSION,
+    HAZELCAST_TEST_VERSION: HAZELCAST_TEST_VERSION,
+    HAZELCAST_ENTERPRISE_KEY: HAZELCAST_ENTERPRISE_KEY,
+    HAZELCAST_ENTERPRISE_VERSION: HAZELCAST_ENTERPRISE_VERSION,
+    HAZELCAST_ENTERPRISE_TEST_VERSION: HAZELCAST_ENTERPRISE_TEST_VERSION
+};
